requirements.txt
Django>=1.8,<1.9
Pillow>=2.8,<3
arrow>=0.4,<1
celery>=3.1,<4
django-absolute>=0.3,<1
django-admin-bootstrapped>=2.5,<3
django-bootstrap3-datetimepicker>=2.2,<3
django-bootstrap3>=5.4,<6
django-endless-pagination>=2.0,<3
django-redis-cache>=1.6,<2
django-redis-sessions>=0.5,<1
django-redis>=4.3,<5
django-registration-redux>=1.2,<2
django-simple-captcha>=0.5,<1
django-stdimage>=1.2,<2
django-websocket-redis>=0.4,<1
git+https://github.com/mdevlamynck/django-multiform.git@pr-api#egg=django-multiform
ics>=0.3,<1
psycopg2>=2.6,<3
redis>=2.10,<3
requests>=2.7,<3
